The includes file includes now all new files.
The contrands file is slightly modified to allow faster debugging.

The auxiliary file now has some stuff added for in the future enabling the use of gnuplot.

The potential function files are update with a new function that is the inverse of the potental function.

The survival function g is updated and more standarised.